backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
psycopg2-binary>=2.9.9
pydantic>=2.5.3
pydantic-settings>=2.1.0
httpx==0.26.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
google-auth>=2.26.2
PyJWT>=2.8.0
sqlalchemy>=2.0.46
alembic==1.13.1
redis==5.0.1
celery==5.3.6
numpy>=1.26.4
pandas>=2.2.0
scikit-learn>=1.4.1
openai>=1.10.0
google-generativeai>=0.4.0
tiktoken>=0.5.2
langchain==0.1.4
langchain-community==0.0.16
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
black==24.1.1
flake8==7.0.0
mypy==1.8.0
